Migrate the FeatureStateStorageWrapper to the togglz-core module to make it usable from other state repositories
